processor/requirements.txt
kafka-python
textblob
